feat: added new components for each tag admin button
