+++ src/pages/monitoring/outlets/tagsAdmin/TagFormBtn.tsx
@@ -29,7 +29,6 @@ import { toast } from "sonner";
 import { PlusIcon, UserRoundCheck } from "lucide-react";
 import {
   Dialog,
-  DialogClose,
   DialogContent,
   DialogDescription,
   DialogFooter,
@@ -40,7 +39,6 @@ import {
 import { translateTagCategory } from "pages/monitoring/outlets/tagsAdmin/utils/tagCategoryTranslator";
 import {
   addTag,
-  deleteTag,
   getTagById,
   getTagCategories,
   updateTag,
@@ -50,7 +48,7 @@ export function TagFormButton({
   value: tagId,
   onActionSuccess,
 }: {
-  value?: unknown;
+  value?: number;
   onActionSuccess?: () => void;
 }) {
   const [errors, setErrors] = useState<Partial<TagDataFormErr>>({});
@@ -59,13 +57,6 @@ export function TagFormButton({
   const [formData, setFormData] = useState<TagDataForm>(makeInitialInfo());
   const [tagCategories, setTagCategories] = useState<TagCategory[]>([]);
   const [openDialogForm, setOpenDialogForm] = useState(false);
-  const [openDialogAlert, setOpenDialogAlert] = useState(false);
-
-  if (tagId != null && typeof tagId !== "number") {
-    throw new Error(
-      `Expected type of value: number, received: ${typeof tagId}`,
-    );
-  }
 
   const fetchTag = async () => {
     if (tagId) {
@@ -108,35 +99,6 @@ export function TagFormButton({
     );
 
     setLoadStatusMsg(uiText.table.loadStatus.loaded);
-  };
-
-  const removeTag = async () => {
-    if (tagId) {
-      setLoadStatusMsg(uiText.table.loadStatus.loading);
-      const res = await deleteTag(tagId);
-      setIsLoading(false);
-
-      if (isMonitoringAPIError(res)) {
-        setErrors((oldErr) => ({
-          ...oldErr,
-          root: res.data.map((error) => error.msg),
-        }));
-        setLoadStatusMsg(uiText.table.loadStatus.loaded);
-        return;
-      }
-
-      setLoadStatusMsg(uiText.table.loadStatus.loaded);
-
-      toast(uiText.toast.delete.title, {
-        position: "bottom-right",
-        description: uiText.toast.delete.description,
-        icon: <UserRoundCheck className="size-8 text-primary" />,
-        className: "px-6! gap-6! border-2! border-primary!",
-      });
-
-      setOpenDialogAlert(false);
-      onActionSuccess?.();
-    }
   };
 
   useEffect(() => {
@@ -473,47 +435,6 @@ export function TagFormButton({
           </DialogDescription>
         </DialogContent>
       </Dialog>
-      {tagId && (
-        <Dialog open={openDialogAlert} onOpenChange={setOpenDialogAlert}>
-          <DialogTrigger asChild>
-            <Button
-              onClick={() => void fetchTag()}
-              disabled={loadStatusMsg !== null}
-              variant="ghost"
-            >
-              {loadStatusMsg !== null
-                ? loadStatusMsg
-                : uiText.table.deleteBtn.defaultText}
-            </Button>
-          </DialogTrigger>
-          <DialogContent className="max-h-[80vh] max-w-[60vh] flex flex-col p-4 md:p-8 overflow-hidden">
-            <div className="pb-2">
-              <DialogHeader>
-                <DialogTitle>{uiText.table.deleteBtn.dialog.title}</DialogTitle>
-              </DialogHeader>
-            </div>
-            <DialogDescription className="grid grid-cols-1 gap-6 [&>p]:m-0 [&>p]:flex [&>p]:flex-col [&>p>span]:first:font-semibold [&>p>span]:first:text-primary">
-              <>
-                {uiText.table.deleteBtn.dialog.description(formData.name)}
-                <ErrorsList
-                  errorItems={errors.root ?? []}
-                  className="bg-red-50 p-4 mt-2 rounded-lg outline-2 outline-accent"
-                />
-              </>
-            </DialogDescription>
-            <DialogFooter>
-              <DialogClose asChild>
-                <Button variant="outline">
-                  {uiText.table.deleteBtn.actionBtns.cancel}
-                </Button>
-              </DialogClose>
-              <Button onClick={() => void removeTag()}>
-                {uiText.table.deleteBtn.actionBtns.confirm}
-              </Button>
-            </DialogFooter>
-          </DialogContent>
-        </Dialog>
-      )}
     </>
   );
 }
